接口改动

+++ src/main/java/ftq/ink/activation/anno/SubscriptionIdValidator.java
@@ -6,13 +6,13 @@ import javax.validation.ConstraintValidatorContext;
 import ftq.ink.activation.dao.SubscriptionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
-public class SubscriptionIdValidator implements ConstraintValidator<ValidSubscriptionId, Integer> {
+public class SubscriptionIdValidator implements ConstraintValidator<ValidSubscriptionId, Long> {
 
     @Autowired
     private SubscriptionRepository subscriptionRepository;
 
     @Override
-    public boolean isValid(Integer subscriptionId, ConstraintValidatorContext context) {
+    public boolean isValid(Long subscriptionId, ConstraintValidatorContext context) {
         // 校验 subscriptionId 是否存在
         return subscriptionRepository != null && subscriptionRepository.existsById(subscriptionId);
     }
