+++ src/main/java/ftq/ink/activation/util/ActivationCodeGenerator.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.UUID;
 
 public class ActivationCodeGenerator {
-    public static List<ActivationCode> batchGeneActivationCode(int subscriptionId, int count) {
+    public static List<ActivationCode> batchGeneActivationCode(Long subscriptionId, int count) {
         List<ActivationCode> codeList = new ArrayList<>();
         for (int i = 0; i < count; i++) {
             ActivationCode activationCode = geneActivationCode(subscriptionId);
@@ -16,7 +16,7 @@ public class ActivationCodeGenerator {
         return codeList;
     }
 
-    public static ActivationCode geneActivationCode(int subscriptionId) {
+    public static ActivationCode geneActivationCode(Long subscriptionId) {
         ActivationCode activationCode = new ActivationCode();
         activationCode.subscriptionId = subscriptionId;
         activationCode.content = UUID.randomUUID().toString().replace("-", "");
